Fixed issue with NijiiroScoring
NijiiroScoring uses ReadCoroutine to read fumen data from the game, by linking the passed in filePath with the resulting __instance.fumenPath. This was previously being assigned to a different value, causing NijiiroScoring to not be able to find the data.
Also NijiiroScoring was calling this function before songsManager was assigned (I assume, as it was null), causing some error messages, but not leading to any real issues.